+++ src/Forms/resources/fields/repeat/repeat.blade.php
@@ -6,8 +6,7 @@
     data-repeat
     data-repeat-endpoint="{{ $getEndpoint() }}"
     data-repeat-section-name="{{ $getName() }}"
-{{--    id="{{ $getElementId($locale ?? null) }}"--}}
-    id="{{ $getId() }}"
+    id="{{ $getElementId($locale ?? null) }}"
     class="space-y-3"
 >
     @foreach($repeatedComponents as $components)