+++ src/Managers/Assistants/CrudAssistant.php
@@ -182,7 +182,7 @@ trait CrudAssistant
      */
     public function create(Request $request)
     {
-        $model = $this->managedModelClassInstance($this->resource->getInstanceAttributes($request));
+        $model = $this->managedModelClassInstance(...$this->resource->getInstanceAttributes($request));
 
         View::share('manager', $this);
         View::share('model', $model);
@@ -214,7 +214,7 @@ trait CrudAssistant
 
     private function handleStore(Request $request)
     {
-        $model = $this->managedModelClassInstance($this->resource->getInstanceAttributes($request));
+        $model = $this->managedModelClassInstance(...$this->resource->getInstanceAttributes($request));
 
         $fields = Forms::make($this->resource->fields($model))
             ->fillModel($model)